Get rid of the dependency on dateutil; output at date granularity
The dateutil library is really nice but can be a pain to install on Windows.
Also modify the timestamps from the FMI data parser to be dates, not full datetimes.
This should make it easier to match the data with the zoo visitor data which are at
the granularity of dates with no times.