Harvest weather forecasts rather than observations
The harvester now retrieves the weather forecast for the next day rather than observations from the previous one.

The daily forecasts are formed by taking hourly forecasts for the day and merging them into daily max, min, mean and total values. Past observations are still retrieved and parsed the old way from daily observation dumps, at least for now.

Also remove the snow depth information from the weather persistence model. Snow depth information is not available in the forecasts so it's not useful as a feature.